# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VAKSINA HR - Kadrlarni boshqaruv dasturi
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: getTasks
      tags: [tasks]
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: assigneeKind, in: query, schema: { type: string, enum: [user, employee] } }
        - { name: assigneeId, in: query, schema: { type: number } }
        - { name: board, in: query, schema: { type: string, enum: [active, all] } }
      responses:
        "200":
          description: Vazifalar ro'yxati
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vazifa"
    post:
      operationId: createTask
      tags: [tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VazifaInput"
      responses:
        "201":
          description: Yaratildi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vazifa"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      parameters:
        - { name: id, in: path, required: true, schema: { type: number } }
      responses:
        "200":
          description: Vazifa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vazifa"
    patch:
      operationId: updateTask
      tags: [tasks]
      parameters:
        - { name: id, in: path, required: true, schema: { type: number } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VazifaUpdate"
      responses:
        "200":
          description: Yangilandi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vazifa"
    delete:
      operationId: deleteTask
      tags: [tasks]
      parameters:
        - { name: id, in: path, required: true, schema: { type: number } }
      responses:
        "204":
          description: O'chirildi

  # --- NOTIFICATIONS ---
...
components:
  schemas:
...
    TaskAttachment:
      type: object
      required: [id, name, mimeType, kind, url]
      properties:
        id: { type: string }
        name: { type: string }
        mimeType: { type: string }
        kind:
          type: string
          enum: [image, file]
        url: { type: string }
        size: { type: number }
    Vazifa:
      type: object
      required: [id, title, status, priority, assigneeKind, assigneeId, createdById, attachments, createdAt, updatedAt]
      properties:
        id: { type: number }
        title: { type: string }
        description: { type: ["string", "null"] }
        status:
          type: string
          enum: [todo, in_progress, done, cancelled]
        priority:
          type: string
          enum: [low, normal, high, urgent]
        dueAt: { type: ["string", "null"] }
        assigneeKind:
          type: string
          enum: [user, employee]
        assigneeId: { type: number }
        assigneeName: { type: ["string", "null"] }
        createdById: { type: number }
        createdByName: { type: ["string", "null"] }
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/TaskAttachment"
        createdAt: { type: string }
        updatedAt: { type: string }
    VazifaInput:
      type: object
      required: [title, assigneeKind, assigneeId]
      properties:
        title: { type: string }
        description: { type: ["string", "null"] }
        status: { type: string }
        priority: { type: string }
        dueAt: { type: ["string", "null"] }
        assigneeKind:
          type: string
          enum: [user, employee]
        assigneeId: { type: number }
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/TaskAttachment"
    VazifaUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: ["string", "null"] }
        status: { type: string }
        priority: { type: string }
        dueAt: { type: ["string", "null"] }
        assigneeKind:
          type: string
          enum: [user, employee]
        assigneeId: { type: number }
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/TaskAttachment"